# Benchmark Json serialization. 25X potential speedup.