--- a/benchmark/paket.references
+++ b/benchmark/paket.references
@@ -2,3 +2,4 @@ group Benchmark
 
 BenchmarkDotNet
 Ply
+Utf8Json