--- a/paket.dependencies
+++ b/paket.dependencies
@@ -29,3 +29,4 @@ group Benchmark
 
     nuget BenchmarkDotNet
     nuget Ply
+    nuget Utf8Json