```
|           Method |      Mean |     Error |    StdDev | Ratio |     Gen 0 |     Gen 1 |    Gen 2 |   Allocated |
|----------------- |----------:|----------:|----------:|------:|----------:|----------:|---------:|------------:|
|             Toth | 59.415 ms | 1.2872 ms | 1.4823 ms |  1.00 | 4555.5556 | 2000.0000 | 777.7778 | 18447.61 KB |
|         Utf8Json |  2.534 ms | 0.0128 ms | 0.0147 ms |  0.04 |  410.1563 |  402.3438 | 332.0313 |  1776.21 KB |
| System.Text.Json |  6.637 ms | 0.0300 ms | 0.0345 ms |  0.11 |  187.5000 |  179.6875 | 125.0000 |   895.16 KB |```

Utf8Json is fastest by a large margin.
System.Text.Json has fewer memory allocations while being 2-3X slower.
Jil was not supported in .NET core 3.1.